refactor: move SetupCoreDump to pkg/util/coredump
