+++ cmd/system-probe/subcommands/run/command.go
@@ -61,7 +61,7 @@ import (
 	ebpftelemetry "github.com/DataDog/datadog-agent/pkg/ebpf/telemetry"
 	processstatsd "github.com/DataDog/datadog-agent/pkg/process/statsd"
 	ddruntime "github.com/DataDog/datadog-agent/pkg/runtime"
-	"github.com/DataDog/datadog-agent/pkg/util"
+	"github.com/DataDog/datadog-agent/pkg/util/coredump"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	pkglog "github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
@@ -333,7 +333,7 @@ func startSystemProbe(log log.Component, statsd compstatsd.Component, telemetry 
 		return ErrNotEnabled
 	}
 
-	if err := util.SetupCoreDump(sysprobeconfig); err != nil {
+	if err := coredump.Setup(sysprobeconfig); err != nil {
 		log.Warnf("cannot setup core dumps: %s, core dumps might not be available after a crash", err)
 	}
