+++ pkg/util/coredump/core_windows.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-package util
+package coredump
 
 import (
 	"fmt"
@@ -11,8 +11,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/config/model"
 )
 
-// SetupCoreDump enables core dumps and sets the core dump size limit based on configuration
-func SetupCoreDump(cfg model.Reader) error {
+// Setup enables core dumps and sets the core dump size limit based on configuration
+func Setup(cfg model.Reader) error {
 	if cfg.GetBool("go_core_dump") {
 		return fmt.Errorf("Not supported on Windows")
 	}
